Finished working on the backend of the application

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("./../models/goalModel");
+const User = require("./../models/userModel");
 
 // @descr: Get all Goals
 //  @Route: GET /api/v1/goals
 // @Access: PRIVATE
 exports.getAllGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json({
     staus: "success",
     result: goals.length,
@@ -27,6 +28,7 @@ exports.createGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.create({
     title: req.body.title,
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(201).json({
@@ -47,6 +49,19 @@ exports.updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  // Check if user exists
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("USer not found");
+  }
+
+  // Checking to see the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(id, req.body, {
@@ -71,6 +86,20 @@ exports.deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  // Check if user exists
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("USer not found");
+  }
+
+  // Checking to see the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
+  }
+
   await Goal.findByIdAndDelete(id);
   res.status(200).json({
     staus: "success",
